Create functions file with reusable functions
